:lipstick: feat: alterando tratativas de respostas da api

+++ src/routes.js
@@ -1,6 +1,6 @@
-import { Database } from './database.js'
-import { buildRoutePath } from './utils/build-route-path.js'
-import { randomUUID } from 'node:crypto'
+import { Database } from './database.js';
+import { buildRoutePath } from './utils/build-route-path.js';
+import { randomUUID } from 'node:crypto';
 
 // Query Parameters: URL stateful => Filtros, paginação, ordenação, group by
 // Ex.: /users?name=rafael
@@ -9,31 +9,37 @@ import { randomUUID } from 'node:crypto'
 // Request Body: Envio de informações para o backend
 // 	Ex.: {"name": "Rafael", "email": "rafael@example.com"}
 
-const database = new Database()
+const database = new Database();
 
 export const routes = [
   {
     method: 'GET',
     path: buildRoutePath('/tasks'),
     handler: (req, res) => {
-      const { search } = req.query
+      const { search } = req.query;
       const tasks = database.select(
         'tasks',
         search
           ? {
               title: search,
-              description: search,
+              description: search
             }
-          : null,
-      )
-      return res.end(JSON.stringify(tasks))
-    },
+          : null
+      );
+      return res.end(JSON.stringify(tasks));
+    }
   },
   {
     method: 'POST',
     path: buildRoutePath('/tasks'),
     handler: (req, res) => {
-      const { title, description, completed_at, created_at, updated_at } = req.body
+      const { title, description } = req.body;
+
+      if (!title || !description) {
+        return res
+          .writeHead(401)
+          .end('Título e descrição são necessários para o correto cadastramento!');
+      }
 
       const task = {
         id: randomUUID(),
@@ -41,53 +47,70 @@ export const routes = [
         description,
         completed_at: null,
         created_at: new Date(),
-        updated_at,
-      }
+        updated_at: null
+      };
 
-      database.insert('tasks', task)
+      database.insert('tasks', task);
 
-      return res.writeHead(201).end()
-    },
+      return res.writeHead(201).end();
+    }
   },
   {
     method: 'PUT',
     path: buildRoutePath('/tasks/:id'),
     handler: (req, res) => {
-      const { id } = req.params
-      const { title, description, completed_at, created_at, updated_at } = req.body
+      const { id } = req.params;
+      const task = database.select('tasks', { id })[0];
+      const { title, description } = req.body;
 
-      database.update('tasks', id, { title, description, completed_at, created_at, updated_at: new Date() })
+      console.log(task, id);
 
-      return res.writeHead(204).end()
-    },
+      if (!task) return res.writeHead(404).end('ID da tarefa não encontrado');
+
+      if (!title || !description) {
+        return res.writeHead(401).end('Título e descrição são necessários para atualizar a task');
+      }
+
+      database.update('tasks', id, { ...task, title, description, updated_at: new Date() });
+
+      return res.writeHead(204).end();
+    }
   },
   {
     method: 'PATCH',
     path: buildRoutePath('/tasks/:id/complete'),
     handler: (req, res) => {
-      const { id } = req.params
+      const { id } = req.params;
 
-      const task = database.select('tasks', { id })[0]
+      const task = database.select('tasks', { id })[0];
 
       if (!task) {
-        return res.writeHead(404).end()
+        return res.writeHead(404).end();
       }
 
-      const updatedTask = { ...task, completed_at: new Date() }
+      const updatedTask = { ...task, completed_at: new Date() };
 
-      database.path('tasks', id, updatedTask)
+      database.path('tasks', id, updatedTask);
 
-      return res.writeHead(204).end()
-    },
+      return res.writeHead(204).end();
+    }
   },
   {
     method: 'DELETE',
     path: buildRoutePath('/tasks/:id'),
     handler: (req, res) => {
-      const { id } = req.params
-      database.delete('tasks', id)
+      const { id } = req.params;
+      const task = database.select('tasks', { id })[0];
 
-      return res.writeHead(204).end()
-    },
-  },
-]
+      if (!task) {
+        return res.writeHead(404).end();
+      }
+
+      console.log(id, { task });
+
+      database.delete('tasks', id);
+
+      return res.writeHead(204).end();
+    }
+  }
+];
